+++ sections/chat.py
@@ -3,13 +3,16 @@ import streamlit as st
 from dotenv import load_dotenv
 from openai import AzureOpenAI
 
+from utils.llm_helpers import setup_azure_openai_client, setup_anthropic_client
+
 # Load environment variables from .env file
 load_dotenv()
 
 
 # Constants
-MODEL_PROVIDERS = ["Azure OpenAI"]
+MODEL_PROVIDERS = ["Azure OpenAI", 'Anthropic']
 AZURE_OPENAI_MODELS = ["gpt-4o-mini", "gpt-4o"]
+ANTHROPIC_MODELS = ["claude-3-5-sonnet-20240620", "claude-3-haiku-20240307"]
 
 
 def initialize_session_state():
@@ -34,28 +37,11 @@ def render_sidebar():
     if model_provider == "Azure OpenAI":
         model_name = st.sidebar.selectbox("Model Name", AZURE_OPENAI_MODELS, key="aoai_model_name")
         st.session_state.model_name = model_name
+    elif model_provider == "Anthropic":
+        model_name = st.sidebar.selectbox("Model Name", ANTHROPIC_MODELS, key="anthropic_model_name")
+        st.session_state.model_name = model_name
 
     st.sidebar.markdown("---")  # Add a line at the end
-
-
-
-def setup_azure_openai_client():
-    """
-    Set up the Azure OpenAI client.
-
-    Returns:
-        AzureOpenAI: The Azure OpenAI client.
-    """
-    try:
-        client = AzureOpenAI(
-            api_key=os.getenv("AZURE_OPENAI_API_KEY"), 
-            api_version="2024-07-01-preview",
-            azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT")
-        )
-        return client
-    except Exception as e:
-        st.error(f"An error occurred while setting up the Azure OpenAI client: {e}")
-        return None
 
 
 def render():
@@ -63,6 +49,7 @@ def render():
     render_sidebar()
     initialize_session_state()
     client = setup_azure_openai_client()
+    client_claude = setup_anthropic_client
 
     for msg in st.session_state.messages:
         st.chat_message(msg["role"]).write(msg["content"])
@@ -72,7 +59,12 @@ def render():
 
         st.session_state.messages.append({"role": "user", "content": prompt})
         st.chat_message("user").write(prompt)
-        response = client.chat.completions.create(model=st.session_state.model_name, messages=st.session_state.messages)
+        if st.session_state.model_provider == "Azure OpenAI":
+            response = client.chat.completions.create(model=st.session_state.model_name, messages=st.session_state.messages)
+        elif st.session_state.model_provider == "Anthropic":
+            response = client_claude.messages.create(model=st.session_state.model_name,
+            messages=[{"role": m["role"], "content": m["content"]}for m in st.session_state.messages]
+            )
         msg = response.choices[0].message.content
         st.session_state.messages.append({"role": "assistant", "content": msg})
         st.chat_message("assistant").write(msg)
